# api/openapi.yaml
openapi: 3.1.0
info:
  title: SDAIA AI Engineer Study Space API
  version: 1.0.0
  description: Future private backend contract. No server is deployed by this repository.
  license: { name: MIT, identifier: MIT }
servers:
  - url: /v1
security: []
components:
  parameters:
    AnonHeader:
      name: X-Anon-Id
      in: header
      required: true
      description: Client-generated anonymous identifier; not authentication.
      schema: { type: string, format: uuid }
    AnonPath:
      name: anon_id
      in: path
      required: true
      schema: { type: string, format: uuid }
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: { type: string }
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema: { $ref: ../data/schema/error.schema.json }
paths:
  /health:
    get:
      operationId: health
      summary: Check service liveness
      responses:
        '200':
          description: Liveness
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties: { status: { type: string, const: ok } }
                additionalProperties: false
  /bank:
    get:
      operationId: getBank
      summary: Load the public runtime bundle
      responses:
        '200':
          description: Canonical public runtime bundle; the active SDAIA track returns RuntimeBundleV4 with K3 evidence context
          content:
            application/json:
              schema: { $ref: ../data/schema/runtime-bundle-v4.schema.json }
  /progress/{anon_id}:
    parameters:
      - { $ref: '#/components/parameters/AnonPath' }
      - { $ref: '#/components/parameters/AnonHeader' }
    get:
      operationId: getProgress
      summary: Load anonymous progress
      description: X-Anon-Id is a client-generated identifier, not authentication, and must match anon_id.
      responses:
        '200':
          description: Saved state
          headers: { ETag: { schema: { type: string } } }
          content:
            application/json:
              schema: { $ref: ../data/schema/state-v2.schema.json }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
    put:
      operationId: putProgress
      summary: Replace anonymous progress
      description: Full replace. X-Anon-Id is not authentication; it and state anon_id must match anon_id. If-Match carries the current numeric version; use 0 for first create.
      parameters:
        - { $ref: '#/components/parameters/IfMatch' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: ../data/schema/state-v2.schema.json }
      responses:
        '200':
          description: Saved state
          headers: { ETag: { schema: { type: string } } }
          content:
            application/json:
              schema: { $ref: ../data/schema/state-v2.schema.json }
        '400': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /feedback:
    post:
      operationId: postFeedback
      summary: Submit anonymous question feedback
      parameters: [{ $ref: '#/components/parameters/AnonHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: ../data/schema/feedback.schema.json }
      responses:
        '200':
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                required: [ok, ref]
                properties: { ok: { type: boolean, const: true }, ref: { type: string } }
                additionalProperties: false
        '400': { $ref: '#/components/responses/Error' }
  /events:
    post:
      operationId: postEvents
      summary: Submit anonymous event batch
      parameters: [{ $ref: '#/components/parameters/AnonHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: ../data/schema/events.schema.json }
      responses:
        '200':
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties: { ok: { type: boolean, const: true } }
                additionalProperties: false
        '400': { $ref: '#/components/responses/Error' }

  /learner-evidence/batch:
    post:
      operationId: postLearnerEvidenceBatch
      summary: Push an authorized learner-evidence batch
      description: Authorization selects the learner principal. X-Anon-Id is not sufficient authorization for learner evidence.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [events]
              additionalProperties: false
              properties:
                events:
                  type: array
                  items: { $ref: ../data/schema/learner-evidence-event-v2.schema.json }
      responses:
        '200':
          description: Per-event evidence acceptance result
          content:
            application/json:
              schema: { $ref: ../data/schema/evidence-batch-result-v1.schema.json }
        '400': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }